Factor out casting implementation for primitive types
